backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
boto3==1.34.14
django-storages==1.14.2
python-decouple==3.8
gunicorn==21.2.0
Pillow==10.2.0
